+++ packages/cli-kit/src/prompt.ts
@@ -1,4 +1,0 @@
-import inquirer from 'inquirer';
-
-const prompt: typeof inquirer = inquirer;
-export default prompt;